Introduces InnerPrivateVariables
